+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ### Fixed
  - Fixed the validation forms to register new workshop for hour and date
  - Fixed the save confirmation with the desicion dialog
+ 
+### Added
+ - Added DecisionDialog Atom
+ - Screen to consult details of external collaborators
+ - Screen to consult list of external collaborators
+ - Screen to register external collaborators
 
 ## [0.1.0] - 9/11/2025